refact(core-settings): spliting folder component
